Orders are now a 3 length item (cmd, pos, target), removed several micro-order functions (see description) and removed duplicate code to check distance to move target
I plan to use micro_orders later when the AI uses pathfinding but I want
to re-write the functions based on the code at the time and not be
constrained by legacy code. Since it's not in use anywhere at the moment
I've stripped it out so I don't start using it by mistake.